requirements.txt
asgiref
Django>=4.2
djangorestframework~=3.14
django-filter
psycopg2-binary
# psycopg2
Pillow
gunicorn

stripe
beautifulsoup4
environs
redis

django-debug-toolbar

django-crispy-forms
crispy-bootstrap5
social-auth-app-django

django-ratelimit
django-redis

# swagger
pyyaml
uritemplate
drf-yasg

